// frontend/awana-app/src/types/event.ts
export interface Event {
  event_ID: number;
  event_Name: string;
  event_Date: string;
  event_Year: number;
  event_Location: string;
  event_Place: string;
  event_Registration_Start_Date: string;
  event_Registration_End_Date: string;
  event_Start_Date: string;
  event_End_Date: string;
  event_Max_Participants: number;
  event_Registration_Fee: number;
  event_Is_Public: boolean;
} 